Major performance improvements for `#additional_component_fields`
Profiling led me to `#additional_component_fields` as the major
source of pain. This (rather too large) pull request has the
following changes:

* DRY up the method to avoid multiple calls to, e.g.,
  `#parent_unittitle_list`
* Use absolute xpaths for document-level data (`eadid`, `repository`,
  and `collection`).
* Remove a spurious round-trip from node to XML and back to a node
  in `#parent_unittitle_list` by using relative xpath calls
* Cache/memoize component data

It's that last point that made it hard to do this in multiple pull
requests.

On large files, the speedups I'm seeing are
significant. On one large EAD (4.5MB, 14K components) processing
time to create a solr document went from 1,438 seconds
to 28 seconds.